PR feedback: - rename all show.go `show*` functions to `print*` - moved getTimestampColAsUint64 to utils - if needing to execute in local mode, check the type of Queryist to figure this out
